+++ esp32c2-hal/examples/spi_halfduplex_read_manufacturer_id.rs
@@ -23,7 +23,7 @@ use esp32c2_hal::{
     peripherals::Peripherals,
     prelude::*,
     spi::{
-        master::{prelude::*, Address, Command, HalfDuplexReadWrite, Spi},
+        master::{Address, Command, HalfDuplexReadWrite, Spi},
         SpiDataMode,
         SpiMode,
     },